Addition of propeller performance map module
